+++ ThirdPartyAdapters/verve/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Verve Android Mediation Adapter Changelog
 
+#### Next Version
+- Filter age-restricted requests for Verve SDK when new `AgeRestrictedTreatment` is set to `CHILD`.
+
 #### Version 3.8.2.0
 - Verified compatibility with Verve HyBid SDK version 3.8.2.